Tested, not great yet but better
