+++ main_test.py
@@ -44,9 +44,10 @@ node = Node(
 
 async def testendpoint():
     await asyncio.sleep(1)
-    node.network_controller.send_message(1,b'',3)
+    node.network_controller.send_message(1,b'TESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTESTTEST',3)
 
-#loop.create_task(loop())
+
+loop.create_task(testendpoint())
 
 
 loop.run_forever()
